feat: Add Terraform GCP configs modified for Org level runners
The original configs for all this were taken from the
`terraform-google-github-actions-runners` that can be found here:
https://github.com/terraform-google-modules/terraform-google-github-actions-runners

However, one core change had to be made to adapt it to our use case:
Changing it from repositry level runners to organization level runners
so they can be used by everyone. This is why we have a `modules/`
directory that references it.